deploy preparation
production and development should use differen query url (localhost / api production url). In order to achieve this, we can put url address in two env files:
- .env.production
- .env.development
so the url will be in process.env.xxx

and in the entry main.vue, we can use mixin to add this data to all components.